For plots on homepage, group vaccination and infection (#59)
In the paper, we combine pre-vaccination/infection and post-vaccination/infection in all the plots, and do not try to draw any distinction between vaccination and infection. But the plots in the repo split out vaccination and infection separately for the main summary heatmap and the `dms-viz`.

This both makes those plots larger and different from how things are split out in the paper.

This pull request merged them to a *Pre vaccination or infection* and *Post vaccination or infection* grouping to correspond to what is used in the paper, and updates the plots and homepage accordingly.

Also stop tracking `sera_group_avgs` in `summaries` as these are old files not generated by anything in `summaries_config` (the current ones are called `sera_group_averages`)